Correct completion return value of initialize IPC server call (#29)
These calls don't return KErrNone to request status but instead returns the handle to the session handler (client fetcher). This returns a placeholder value that is not negative (0x400000), to fit the fact that every function checks if the handle is lower or equal to zero.

This make game survives to game window creation and request server to read a custom command buffer, that contains another opcode for device creation, clear, read/write pixels.