[loader] Fix missing exception relocations
This was caused by a calculation: since the offset we start to fill in relocation is zero, after we filling x relocations, the iterator should increase by x - 1 (counting zero element also). The bug caused was missing a one subtract, causing the last relocations set to be missing and data remains unrelocate.