[loader] Not counting reserved relocation as a "real" relocation
The last relocation set was discovered by not counting relocation code 0 (reserved) as a real relocation. By that, if there is a 0 code, subtract iterator by 1. This works, and reveal the last relocation set. Since relocation for CPP relocations address are mostly stays at the end, this fixes many crash when exception happens at many servers.