arm/llvm: Load registers and store registers
Added two value arrays: local and globals. Local indicates local change to the register and etc.., that will be written during flushing.

Globals are reference to ARM context's variables. If a index in globals is not null, that means a write has happen, so in that case, local changes are appiled to the context's value.