intests: Add bitmap zero size test
Turns out it's valid after all.
I didn't include one test in here since it's not really valid if some process is also creating a bitmap, but zero bitmap is like a placeholder. Nothing is allocated though. It's prove by created two bitmaps really quick and compare the pointer. Apprently, it's also a fact that the top of the allocator pointer is taken for the data offset. I will note this to myself.