epoc/kernel/libmngr: Add ability to patch ROM dlls with another one
Load the provided DLL into ROM section, then try the best to patch original code in ROM to trampoline to the new export function.

You can map various exports with a .map file, with same name as the DLL you want to patch

Each line of the file will have the formula of: <source_export>,<dest_export> where source_export is the ordinal of the export you want to patch in ROM image, and dest_export is the ordinal of the export you want to replace the original one.