services/window/scheduler: Correct the screen lock not being in right place
when session_send is being executed, the kernel is locked. Therefore, this draw operation is locked when the kernel lock is trying to accquired. But if we lock the screen before kernel lock, then delivering other events to window server is also locked up too, and deliver events is one part of the graphics thread, which session_send may depend too. This fix this circular lock dependency