services/window: Simplifiy client-use drawing code, implement visible region
Fix the part of the architecture issue where the previous window can still be render despite being hidden. This is too clear when a window draws it background, that visible region has to be brought into the place.

Of course, this is far from finish, since theres no clipping method to clip the visible region and draw it out. If there is any more problem, we should take up from this commit.

Plus transparent region is not implemented (region where even parent window overlapped it, still visible)